+++ src/archive/rar.rs
@@ -14,12 +14,7 @@ use crate::{
 
 /// Unpacks the archive given by `archive_path` into the folder given by `output_folder`.
 /// Assumes that output_folder is empty
-pub fn unpack_archive(
-    archive_path: &Path,
-    output_folder: &Path,
-    password: Option<&[u8]>,
-    quiet: bool,
-) -> crate::Result<Unpacked> {
+pub fn unpack_archive(archive_path: &Path, output_folder: &Path, password: Option<&[u8]>) -> crate::Result<Unpacked> {
     let archive = match password {
         Some(password) => Archive::with_password(archive_path, password),
         None => Archive::new(archive_path),
@@ -31,13 +26,11 @@ pub fn unpack_archive(
     while let Some(header) = archive.read_header()? {
         let entry = header.entry();
         archive = if entry.is_file() {
-            if !quiet {
-                info!(
-                    "extracted ({}) {}",
-                    Bytes::new(entry.unpacked_size),
-                    entry.filename.display(),
-                );
-            }
+            info!(
+                "extracted ({}) {}",
+                Bytes::new(entry.unpacked_size),
+                entry.filename.display(),
+            );
             unpacked += 1;
             header.extract_with_base(output_folder)?
         } else {